cxgb4: Clear connection resources when pf4 device is removed
As ch_ktls and ch_ipsec dev_ops are registered with cxgb4, when pf4
device is removed, the uld_ctx pointer becomes null and the call to
driver's dev_opp functions is missed. Due to this the connection
resources remain active even after removal of pf4 device.

To avoid this added a condition in chcr_offload_state(), which checks
if device is shutting down and dev_ops pointer is valid then call the
driver's(ch_ipsec, ch_ktls) dev_ops functions.

Signed-off-by: Ayush Sawal <ayush.sawal@chelsio.com>