ch_ktls: bug 40962: fix kernel panic
offload Tx queues is shared among ULDs and initialized by
ULD registered first. ch_ktls driver is wrongly initializing
number of tx queue to zero, which is causing kernel panic
when cxgbit ULD is operating on it.
fix it by initializing the number of tx queue for offload
Tx queue.

Signed-off-by: Vinay Kumar Yadav <vinay.yadav@chelsio.com>